Revert move of .flat from frame > border to frame
Changing code to agree with docs, which said frame.flat, was backwards.
Mea culpa. Theme authors ran with the actual behaviour, not the docs. As
stability is more important, let’s go back to frame > border.flat, and
fix the docs to reflect what the code does and how to set .flat in code.

N.B. This drops the > from the frame border selector in the CSS files,
because in the patches being reverted here, that was only changed in the
generated CSS files, not the source SCSS. It is more correct not to
include that change in the CSS files until it reaches their source SCSS.

https://bugzilla.gnome.org/show_bug.cgi?id=778905

Gbp-Pq: Name Revert-move-of-.flat-from-frame-border-to-frame.patch