gdk: Don't distribute generated files in tarballs
gdkversionmacros.h is generated by configure, so it should be in
DISTCLEANFILES. The rest of $(gdk_built_sources) are built at compile
time by GLib tools, so in principle they should be in CLEANFILES, but
DISTCLEANFILES is close enough.

This is a continuation of the same general topic as commit dad773b1.

Signed-off-by: Simon McVittie <smcv@debian.org>
Forwarded: https://gitlab.gnome.org/GNOME/gtk/merge_requests/1001
Bug: https://gitlab.gnome.org/GNOME/gtk/issues/357

Gbp-Pq: Name gdk-Don-t-distribute-generated-files-in-tarballs.patch